+++ docs/versionhistory.rst
@@ -12,6 +12,10 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Fixed the ``BrokenResourceError`` raised by the asyncio ``SocketStream`` not having
   the original exception as its cause
   (`#1055 <https://github.com/agronholm/anyio/issues/1055>`_; PR by @veeceey)
+- Fixed ``TaskGroup`` raising a cancellation exception on asyncio in a level-cancelled
+  scope or an edge-cancelled task when none of its children raised a cancellation
+  exception (`#1069 <https://github.com/agronholm/anyio/issues/1069>`_; PR by
+  @gschaffner)
 
 **4.12.1**